dev_requirements.txt
ipdb==0.11
django-debug-toolbar==1.9.1
pylama==7.4.3
pylint==1.8.2
flake8==3.5.0
pydocstyle==2.1.1
